+++ CHANGELOG.md
@@ -14,6 +14,8 @@ All notable changes to this library are documented in this file.
 
 - Add `_check_dueling_tropics` VTEC check looking for TR+HU overlap (#930).
 - Cross check WCNs against watch database storage for PDS status (#925).
+- For `iemapp` if the `memcachekey=` callable returns `None`, the check of
+memcache is short circuited.
 
 ### Bug Fixes